fix(test): Remove certain characters from test names to avoid ci failing
For whatever reason Kotest needs to create a file with the name of each test,
but of course there are restrictions on what characters can be in filenames.
Even better, those restrictions are wildly different between operating systems.
It's entirely possible that a test named “COM” is unable to output its results
on Windows (because that's an illegal filename on Windows).
I've decided that I don't even want to know if this is the case.

This behavior is stupid. Kotest not catching it beforehand is even more stupid.